+++ frontend/src/Components/Session/Partials/BeltExpectation.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from 'react-i18next';
 // AA Belt Radar
 import type { Session } from "@/Api/schema";
 import { renderTooltip } from '@/Components/Helpers/functions';
-export function SessionBeltExpectation({ sessionData }: { sessionData: Session}) {
+export function SessionBeltExpectation({ sessionData }: { sessionData?: Session}) {
     const { t } = useTranslation();
     return (
         <div className="row text-muted small mt-2">